v1.0.0-cleanup reorganized some stuff, and moved some values to config.ts

+++ src/events/mediaChannels.ts
@@ -1,21 +1,7 @@
 import { EventModule } from "../handler";
 import { Events, Message, PermissionFlagsBits } from "discord.js";
 import Logger from "../handler/util/Logger";
-
-const emojiSets = [
-    ["👍", "😂", "😳", "🤔", "😲", "😢", "😡", "🧵"],
-    ["❤️", "🤩", "🤯", "😳", "🤨", "😐", "😢", "🧵"],
-    ["👍", "👎", "❔", "🧵"],
-];
-
-const config = {
-  "1241695709212704778": { emojiSet: emojiSets[0], textAllowed: false }, // media
-  "1241669747108483072": { emojiSet: emojiSets[0], textAllowed: false }, // creations
-  "1241444035512500295": { emojiSet: emojiSets[1], textAllowed: false }, // art
-  "1241444691015110778": { emojiSet: emojiSets[0], textAllowed: false }, // memes
-  "1241431269527523409": { emojiSet: emojiSets[2], textAllowed: true }, // gamenight-suggestions
-  "1241706024969306182": { emojiSet: emojiSets[0], textAllowed: false }, // gamenight-media
-}
+import { mediaChannels } from "../config";
 
 export = {
     name: Events.MessageCreate,
@@ -23,7 +9,7 @@ export = {
         if (!message?.guildId) return;
 
         const channelId = String(message.channelId);
-        const channelConfig = (config as Record<string, { emojiSet: string[], textAllowed: boolean }>)[channelId];
+        const channelConfig = mediaChannels[channelId];
         if (!channelConfig) return;
 
         if (message.author.id === message.client.user.id) {
@@ -31,10 +17,12 @@ export = {
           return;
         }
 
-        const isMemberImmune = message.member?.permissions.has(PermissionFlagsBits.ManageMessages, true)
+        const isMemberImmune = message.member?.permissions.has(PermissionFlagsBits.ManageMessages, true);
+        const isMemberBooster = message.member?.premiumSince;
 
         if (message.attachments.size === 0 && message.embeds.length === 0 && !channelConfig.textAllowed) {
           if (isMemberImmune) return;
+          if (isMemberBooster) return;
         
           message.delete();
           return;
